requirements.txt
aiofiles==24.1.0
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.7.9
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
distro==1.9.0
fastapi==0.116.1
greenlet==3.2.3
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.10
jiter==0.10.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.26
langchain-core==0.3.68
langchain-openai==0.3.27
langchain-text-splitters==0.3.8
langgraph==0.5.2
langgraph-checkpoint==2.1.0
langgraph-prebuilt==0.5.2
langgraph-sdk==0.1.72
langsmith==0.4.5
openai==1.95.1
orjson==3.10.18
ormsgpack==1.10.0
packaging==24.2
pydantic==2.11.7
pydantic_core==2.33.2
PyPDF2==3.0.1
python-dotenv==1.1.1
python-multipart==0.0.20
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.4
requests-toolbelt==1.0.0
sniffio==1.3.1
SQLAlchemy==2.0.41
starlette==0.47.1
tenacity==9.1.2
tiktoken==0.9.0
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
uvicorn==0.35.0
watchfiles==1.1.0
websockets==15.0.1
xxhash==3.5.0
zstandard==0.23.0

# Graph visualization dependencies
pillow==10.1.0
matplotlib==3.8.2
